Node: Fix security code review issues and update node ID data type
1. QNN: Fix potential file handle leak in CreateFromBinaryFile by ensuring fclose is called.
2. Config: Update node ID parsing to use uint8_t instead of uint32_t across Camera,
DepthFromStereo, OpticalFlow, QNN, Radar, Remap, and Voxelization nodes to match definition.
3. Remap/Voxelization: Reset object state to QC_OBJECT_STATE_INITIAL in DeInitialize.

+++ source/Node/Radar/Radar.cpp
@@ -99,7 +99,7 @@ QCStatus_e RadarConfigIfs::ParseStaticConfig( DataTree &dt, std::string &errors 
     if ( QC_STATUS_OK == status )
     {
         m_config.nodeId.name = dt.Get<std::string>( "name", "" );
-        m_config.nodeId.id = dt.Get<uint32_t>( "id", UINT32_MAX );
+        m_config.nodeId.id = dt.Get<uint8_t>( "id", UINT8_MAX );
 
         m_config.numOfEntries = 2;
